requirements.txt
appium-python-client>=5.2.5
selenium>=4.14.0,<4.37.0
pytest>=7.4.0,<8.0
pytesseract
pillow
openai
python-dotenv

# Also install this when appium is installed:
# appium driver install uiautomator2
